All link functionalities working (with some concerns)
There are some problems with order of execution but with current settings this seems to work. Some lines of code (seemingly unimportant) are potentially delaying the code just long enough for some functions to catch up.

I am worried that this may pose problems on other machines and behave unperdicitbly